simulator/pyproject.toml
[project]
name = "camera-simulator"
version = "0.1.0"
description = "Reolink + Dahua camera simulator for development and E2E testing"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9",
    "pycryptodome>=3.20",
    "av>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
